rename legendItem to legendLayer and split components more
